requirements.txt
flask
flask_socketio
openai-whisper